--- libs/core/src/components/breadcrumb/breadcrumb.ts
@@ -1,5 +1,0 @@
-import { GdsBreadcrumb } from './breadcrumb.component'
-
-GdsBreadcrumb.define()
-
-export { GdsBreadcrumb }
