--- libs/core/src/components/breadcrumb/index.ts
@@ -1,1 +1,0 @@
-export * from './breadcrumb'
