+++ libs/core/src/components/breadcrumbs/breadcrumbs.stories.ts
@@ -2,7 +2,7 @@ import { html } from 'lit'
 
 import type { Meta, StoryObj } from '@storybook/web-components'
 
-import './breadcrumb'
+import './breadcrumbs'
 import '../link'
 import '../flex'
 import '../card'
@@ -20,33 +20,33 @@ import '../icon/icons/folder'
 
 /**
  *
- * The `gds-breadcrumb` component is a navigation element <br/>
+ * The `gds-breadcrumbs` component is a navigation element <br/>
  * It helps users understand their current * location within a hierarchical structure of a website or application.
  *
  * ## Features
  * - Supports both text-only and icon+text navigation links
  * - Responsive design with mobile optimization showing the previous page
  * - Configurable sizes (large/small)
- * - Each breadcrumb item is a slot that preferably uses `gds-link` component.
+ * - Each breadcrumbs item is a slot that preferably uses `gds-link` component.
  *
  * ## Basic Usage
  *
- * ### Simple Breadcrumb
+ * ### Simple breadcrumbs
  * The most basic implementation using text-only links.
  *
  * ```html
- * <gds-breadcrumb>
+ * <gds-breadcrumbs>
  *   <gds-link href="/">Home</gds-link>
  *   <gds-link href="/products">Products</gds-link>
  *   <gds-link href="/products/category">Category</gds-link>
  *   <gds-text>Current page</gds-text>
- * </gds-breadcrumb>
+ * </gds-breadcrumbs>
  * ```
  */
 
 const meta: Meta = {
-  title: 'Components/Breadcrumb',
-  component: 'gds-breadcrumb',
+  title: 'Components/Breadcrumbs',
+  component: 'gds-breadcrumbs',
   tags: ['autodocs'],
 }
 
@@ -55,33 +55,33 @@ type Story = StoryObj
 
 export const Simple: Story = {
   render: () => html`
-    <gds-breadcrumb>
+    <gds-breadcrumbs>
       <gds-link href="/">Home</gds-link>
       <gds-link href="/products">Products</gds-link>
       <gds-link href="/products/category">Category</gds-link>
       <gds-text>Current page</gds-text>
-    </gds-breadcrumb>
+    </gds-breadcrumbs>
   `,
 }
 
 /**
- * This example shows how to use icons alongside text in the breadcrumb links.
+ * This example shows how to use icons alongside text in the breadcrumbs links.
  * Icons can be used to enhance the visual representation of the links.
  *
  * ```html
- * <gds-breadcrumb>
+ * <gds-breadcrumbs>
  *  <gds-link href="/">
  *   <gds-icon-home-open slot="lead"></gds-icon-home-open>
  *   Home
  *  </gds-link>
- * </gds-breadcrumb>
+ * </gds-breadcrumbs>
  * ```
  *
  * For more information on lead and trail icon, please refer to the [Link documentation](https://storybook.seb.io/latest/core/?path=/docs/components-link--docs).
  */
 export const WithIcon: Story = {
   render: () => html`
-    <gds-breadcrumb>
+    <gds-breadcrumbs>
       <gds-link href="/">
         <gds-icon-home-open slot="lead"></gds-icon-home-open>
         Home
@@ -98,16 +98,16 @@ export const WithIcon: Story = {
       </gds-link>
 
       <gds-text> Current page </gds-text>
-    </gds-breadcrumb>
+    </gds-breadcrumbs>
   `,
 }
 
 /**
- * Size of the breadcrumb can be adjusted using the `size` attribute. <br/>
+ * Size of the breadcrumbs can be adjusted using the `size` attribute. <br/>
  * The default size is `large`, but you can set it to `small` for a more compact appearance.
  *
  * ```html
- * <gds-breadcrumb size="small">...</gds-breadcrumb>
+ * <gds-breadcrumbs size="small">...</gds-breadcrumbs>
  * ```
  */
 export const Size: Story = {
@@ -117,7 +117,7 @@ export const Size: Story = {
         <gds-flex align-items="center" gap="m">
           <gds-text width="200px">Default(large)</gds-text>
           <gds-card flex="1" background="secondary">
-            <gds-breadcrumb>
+            <gds-breadcrumbs>
               <gds-link href="/home">
                 <gds-icon-home-open slot="lead"></gds-icon-home-open>
                 Home
@@ -134,14 +134,14 @@ export const Size: Story = {
               </gds-link>
 
               <gds-text> Current page </gds-text>
-            </gds-breadcrumb>
+            </gds-breadcrumbs>
           </gds-card>
         </gds-flex>
 
         <gds-flex align-items="center" gap="m">
           <gds-text width="200px">Small</gds-text>
           <gds-card background="secondary" flex="1">
-            <gds-breadcrumb size="small">
+            <gds-breadcrumbs size="small">
               <gds-link href="/home">
                 <gds-icon-home-open slot="lead"></gds-icon-home-open>
                 Home
@@ -158,7 +158,7 @@ export const Size: Story = {
               </gds-link>
 
               <gds-text> Current page </gds-text>
-            </gds-breadcrumb>
+            </gds-breadcrumbs>
           </gds-card>
         </gds-flex>
       </gds-flex>
@@ -167,8 +167,8 @@ export const Size: Story = {
 }
 
 /**
- * Mobile optimization is built into the breadcrumb component. <br/>
- * On mobile devices, the breadcrumb will show the previous page instead of the full path. <br/>
+ * Mobile optimization is built into the breadcrumbs component. <br/>
+ * On mobile devices, the breadcrumbs will show the previous page instead of the full path. <br/>
  * This is done to save space and make navigation easier on smaller screens.
  */
 export const Mobile: Story = {
@@ -178,7 +178,7 @@ export const Mobile: Story = {
         <gds-flex align-items="center" gap="m">
           <gds-text width="200px">Desktop</gds-text>
           <gds-card flex="1" background="secondary">
-            <gds-breadcrumb>
+            <gds-breadcrumbs>
               <gds-link href="/home">
                 <gds-icon-home-open slot="lead"></gds-icon-home-open>
                 Home
@@ -195,14 +195,14 @@ export const Mobile: Story = {
               </gds-link>
 
               <gds-text> Current page </gds-text>
-            </gds-breadcrumb>
+            </gds-breadcrumbs>
           </gds-card>
         </gds-flex>
 
         <gds-flex align-items="center" gap="m">
           <gds-text width="200px">Mobile</gds-text>
           <gds-card background="secondary" width="400px">
-            <gds-breadcrumb>
+            <gds-breadcrumbs>
               <gds-link href="/home">
                 <gds-icon-home-open slot="lead"></gds-icon-home-open>
                 Home
@@ -219,7 +219,7 @@ export const Mobile: Story = {
               </gds-link>
 
               <gds-text> Current page </gds-text>
-            </gds-breadcrumb>
+            </gds-breadcrumbs>
           </gds-card>
         </gds-flex>
       </gds-flex>
@@ -228,24 +228,24 @@ export const Mobile: Story = {
 }
 
 /**
- * The breadcrumb component is designed with flexibility in mind, allowing developers to implement * custom overflow solutions that best fit their specific needs.
+ * The breadcrumbs component is designed with flexibility in mind, allowing developers to implement * custom overflow solutions that best fit their specific needs.
  *
  * ### Example
  *
  *```html
- *<gds-breadcrumb>
+ *<gds-breadcrumbs>
  *  <gds-link href="/home">Home</gds-link>
  *  <gds-button rank="tertiary">
  *    <gds-icon-dot-grid-one-horizontal></gds-icon-dot-grid-one-horizontal>
  *  </gds-button>
  *  <gds-link href="/current">Current</gds-link>
- *</gds-breadcrumb>
+ *</gds-breadcrumbs>
  *```
  */
 export const Overflow: Story = {
   render: () => html`
     <gds-flex flex-direction="column" gap="4xl">
-      <gds-breadcrumb>
+      <gds-breadcrumbs>
         <gds-link href="/home">
           <gds-icon-home-open slot="lead"></gds-icon-home-open>
           Home
@@ -266,20 +266,20 @@ export const Overflow: Story = {
         </gds-link>
 
         <gds-text> Current page </gds-text>
-      </gds-breadcrumb>
+      </gds-breadcrumbs>
     </gds-flex>
   `,
 }
 
 /**
- * The breadcrumb component supports customizing the accessibility label. <br/>
+ * The breadcrumbs component supports customizing the accessibility label. <br/>
  * This can be useful when you have multiple navigation sections
  * or want to provide more specific context for screen readers.
  *
  * ```html
- * <gds-breadcrumb label="Site Navigation">
+ * <gds-breadcrumbs label="Site Navigation">
  *   ...
- * </gds-breadcrumb>
+ * </gds-breadcrumbs>
  * ```
  */
 export const Label: Story = {
@@ -290,28 +290,28 @@ export const Label: Story = {
         <gds-flex align-items="center" gap="m">
           <gds-text width="200px">Default label</gds-text>
           <gds-card flex="1" background="secondary">
-            <gds-breadcrumb>
+            <gds-breadcrumbs>
               <gds-link href="/home">
                 <gds-icon-home-open slot="lead"></gds-icon-home-open>
                 Home
               </gds-link>
               <gds-link href="/products">Products</gds-link>
               <gds-text>Current page</gds-text>
-            </gds-breadcrumb>
+            </gds-breadcrumbs>
           </gds-card>
         </gds-flex>
 
         <gds-flex align-items="center" gap="m">
           <gds-text width="200px">Custom label</gds-text>
           <gds-card background="secondary" flex="1">
-            <gds-breadcrumb label="Site Navigation">
+            <gds-breadcrumbs label="Site Navigation">
               <gds-link href="/home">
                 <gds-icon-home-open slot="lead"></gds-icon-home-open>
                 Home
               </gds-link>
               <gds-link href="/products">Products</gds-link>
               <gds-text>Current page</gds-text>
-            </gds-breadcrumb>
+            </gds-breadcrumbs>
           </gds-card>
         </gds-flex>
       </gds-flex>
