+++ libs/core/src/components/breadcrumbs/breadcrumbs.ts
@@ -1,0 +1,5 @@
+import { GdsBreadcrumbs } from './breadcrumbs.component'
+
+GdsBreadcrumbs.define()
+
+export { GdsBreadcrumbs }
