+++ libs/core/src/components/breadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export * from './breadcrumbs'
