+++ libs/core/src/components/index.ts
@@ -3,7 +3,7 @@ export * from '../primitives/menu/menu-heading'
 export * from '../primitives/menu/menu-item'
 export * from './button'
 export * from './badge'
-export * from './breadcrumb'
+export * from './breadcrumbs'
 export * from './menu-button'
 export * from './context-menu'
 export * from './coachmark'
