+++ libs/core/src/components/pure.ts
@@ -3,7 +3,7 @@ export * from '../primitives/menu/menu-heading.component'
 export * from '../primitives/menu/menu-item.component'
 export * from './button/button.component'
 export * from './badge/badge.component'
-export * from './breadcrumb/breadcrumb.component'
+export * from './breadcrumbs/breadcrumbs.component'
 export * from './menu-button/menu-button.component'
 export * from './context-menu/context-menu.component'
 export * from './coachmark/coachmark.component'
